feat: Display Metadata of pictures in the slider
Added an AsyncTaskLoader that perform requests to the Flickr API
to get the metadata of the current displayed photo and update
the view.

+++ app/src/androidTest/java/me/oueslati/fakher/flickrapp/util/FlickrJsonInstrumentedTest.java
@@ -76,7 +76,7 @@ public class FlickrJsonInstrumentedTest {
     public void getPhotoInfoFromJson_isCorrect() throws Exception {
         Photo photo = new Photo();
         Photo result = FlickrJsonUtil.getPhotoInfoFromJson(photo, photoInfoJsonData);
-        int expectedDate = 1487348602;
+        long expectedDate = 1487348602;
         String expectedRealname = "Lou Feltz";
         String expectedrofilePictureURL = "http://farm8.staticflickr.com/" +
                 "7446/buddyicons/42415395@N07.jpg";
